Split out TextEncoder
