+++ src/rewritable_units/mod.rs
@@ -1,10 +1,13 @@
 use std::any::Any;
 
+pub(crate) use self::mutations::{Mutations, StringChunk};
+pub(crate) use self::text_encoder::{IncompleteUtf8Resync, TextEncoder};
+
 pub use self::document_end::*;
 pub use self::element::*;
 pub use self::mutations::{ContentType, StreamingHandler};
-pub(crate) use self::mutations::{Mutations, StringChunk};
-pub use self::text_encoder::{StreamingHandlerSink, Utf8Error};
+pub use self::streaming_sink::StreamingHandlerSink;
+pub use self::text_encoder::Utf8Error;
 pub use self::tokens::*;
 
 /// Data that can be attached to a rewritable unit by a user and shared between content handler
@@ -85,6 +88,7 @@ mod mutations;
 
 mod document_end;
 mod element;
+mod streaming_sink;
 mod text_encoder;
 mod tokens;
